cleanup unused / deprecated code
