+++ lib/screens/subsettings/export_import/export_column_data.dart
@@ -1,12 +1,14 @@
+/* TODO: rewrite
+
 import 'package:blood_pressure_app/components/consistent_future_builder.dart';
 import 'package:blood_pressure_app/model/blood_pressure.dart';
-import 'package:blood_pressure_app/model/export_import/legacy_column.dart';
+import 'package:blood_pressure_app/model/export_import/import_field_type.dart';
 import 'package:blood_pressure_app/model/export_options.dart';
 import 'package:blood_pressure_app/screens/subsettings/export_import/export_field_format_documentation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
-class  EditExportColumnPage extends StatefulWidget {
+class EditExportColumnPage extends StatefulWidget {
   final String? initialInternalName;
   final String? initialDisplayName;
   final String? initialFormatPattern;
@@ -191,3 +193,5 @@ class _EditExportColumnPageState extends State<EditExportColumnPage> {
     );
   }
 }
+
+ */
