+++ test/ui/navigation_test.dart
@@ -2,7 +2,6 @@ import 'package:blood_pressure_app/components/dialoges/add_measurement.dart';
 import 'package:blood_pressure_app/components/dialoges/enter_timeformat.dart';
 import 'package:blood_pressure_app/main.dart';
 import 'package:blood_pressure_app/model/blood_pressure.dart';
-import 'package:blood_pressure_app/model/export_import/legacy_column.dart';
 import 'package:blood_pressure_app/model/ram_only_implementations.dart';
 import 'package:blood_pressure_app/model/storage/db/config_dao.dart';
 import 'package:blood_pressure_app/model/storage/export_columns_store.dart';
@@ -20,7 +19,6 @@ import 'package:provider/provider.dart';
 void main() {
   group('start page', () {
     testWidgets('should navigate to add measurements page', (widgetTester) async {
-      globalConfigDao = MockConfigDao();
       await pumpAppRoot(widgetTester);
       expect(find.byIcon(Icons.add), findsOneWidget);
       await widgetTester.tap(find.byIcon(Icons.add));
@@ -91,16 +89,8 @@ Future<void> pumpAppRoot(WidgetTester widgetTester, {
 }
 
 class MockConfigDao implements ConfigDao {
-  Map<String, LegacyExportColumn> columns = {};
-
-  @override
-  Future<void> deleteExportColumn(String internalName) async => columns.remove(internalName);
-
   @override
   Future<CsvExportSettings> loadCsvExportSettings(int profileID) async => CsvExportSettings();
-
-  @override
-  Future<List<LegacyExportColumn>> loadExportColumns() async => columns.values.toList();
 
   @override
   Future<ExportSettings> loadExportSettings(int profileID) async => ExportSettings();
@@ -114,12 +104,7 @@ class MockConfigDao implements ConfigDao {
   @override
   Future<Settings> loadSettings(int profileID) async => Settings();
 
-  @override
-  Future<void> updateExportColumn(LegacyExportColumn exportColumn) async => columns.update(exportColumn.internalName, (value) => exportColumn);
-
-  void reset() {
-    columns = {};
-  }
+  void reset() {}
 
   @override
   Future<ExportColumnsManager> loadExportColumnsManager(int profileID) async => ExportColumnsManager();
